Fix up pretty printer for bicep params files
