+++ src/Bicep.Core.Samples/Files/baselines_bicepparam/Invalid_Parameters/parameters.formatted.bicepparam
@@ -1,4 +1,4 @@
-parampara1= 'value
+param para1 =   'value
 
 para
